Simplify formatter tests
Self-contained test cases. More verbose but clearer.
